requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
asyncpg
requests
python-dotenv